+++ GUI/GUI.csproj
@@ -4,7 +4,7 @@
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <ProjectGuid>{B0AD8FFB-5DA9-44FD-96B9-B3666FDB1917}</ProjectGuid>
+    <ProjectGuid>{4A9F7FB1-E6DD-4789-8F0A-153B862FD38D}</ProjectGuid>
     <OutputType>WinExe</OutputType>
     <RootNamespace>GUI</RootNamespace>
     <AssemblyName>GUI</AssemblyName>
@@ -33,10 +33,17 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="CommonServiceLocator, Version=2.0.4.0, Culture=neutral, PublicKeyToken=489b6accfaf20ef0, processorArchitecture=MSIL">
+      <HintPath>..\packages\Unity.5.8.11\lib\net46\CommonServiceLocator.dll</HintPath>
+    </Reference>
     <Reference Include="ObjectListView, Version=2.9.1.1072, Culture=neutral, PublicKeyToken=b1c5bf581481bcd4, processorArchitecture=MSIL">
       <HintPath>..\packages\ObjectListView.Official.2.9.1\lib\net20\ObjectListView.dll</HintPath>
     </Reference>
     <Reference Include="System" />
+    <Reference Include="System.Configuration" />
+    <Reference Include="System.Configuration.ConfigurationManager, Version=4.0.0.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
+      <HintPath>..\packages\System.Configuration.ConfigurationManager.4.4.1\lib\net461\System.Configuration.ConfigurationManager.dll</HintPath>
+    </Reference>
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -47,6 +54,27 @@
     <Reference Include="System.Net.Http" />
     <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml" />
+    <Reference Include="Unity.Abstractions, Version=3.3.1.0, Culture=neutral, PublicKeyToken=6d32ff45e0ccc69f, processorArchitecture=MSIL">
+      <HintPath>..\packages\Unity.5.8.11\lib\net46\Unity.Abstractions.dll</HintPath>
+    </Reference>
+    <Reference Include="Unity.Configuration, Version=5.2.5.0, Culture=neutral, PublicKeyToken=6d32ff45e0ccc69f, processorArchitecture=MSIL">
+      <HintPath>..\packages\Unity.5.8.11\lib\net46\Unity.Configuration.dll</HintPath>
+    </Reference>
+    <Reference Include="Unity.Container, Version=5.8.11.0, Culture=neutral, PublicKeyToken=489b6accfaf20ef0, processorArchitecture=MSIL">
+      <HintPath>..\packages\Unity.5.8.11\lib\net46\Unity.Container.dll</HintPath>
+    </Reference>
+    <Reference Include="Unity.Interception, Version=5.5.5.0, Culture=neutral, PublicKeyToken=489b6accfaf20ef0, processorArchitecture=MSIL">
+      <HintPath>..\packages\Unity.5.8.11\lib\net46\Unity.Interception.dll</HintPath>
+    </Reference>
+    <Reference Include="Unity.Interception.Configuration, Version=5.1.7.0, Culture=neutral, PublicKeyToken=489b6accfaf20ef0, processorArchitecture=MSIL">
+      <HintPath>..\packages\Unity.5.8.11\lib\net46\Unity.Interception.Configuration.dll</HintPath>
+    </Reference>
+    <Reference Include="Unity.RegistrationByConvention, Version=2.1.9.0, Culture=neutral, PublicKeyToken=489b6accfaf20ef0, processorArchitecture=MSIL">
+      <HintPath>..\packages\Unity.5.8.11\lib\net46\Unity.RegistrationByConvention.dll</HintPath>
+    </Reference>
+    <Reference Include="Unity.ServiceLocation, Version=2.1.2.0, Culture=neutral, PublicKeyToken=489b6accfaf20ef0, processorArchitecture=MSIL">
+      <HintPath>..\packages\Unity.5.8.11\lib\net46\Unity.ServiceLocation.dll</HintPath>
+    </Reference>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="DragControl.cs">
@@ -60,11 +88,17 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="UserControl\UserControlHelp.cs">
+    <Compile Include="UserControl\UserControl1Test.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="UserControl\UserControlHelp.Designer.cs">
-      <DependentUpon>UserControlHelp.cs</DependentUpon>
+    <Compile Include="UserControl\UserControl1Test.Designer.cs">
+      <DependentUpon>UserControl1Test.cs</DependentUpon>
+    </Compile>
+    <Compile Include="UserControl\UserControlCases.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="UserControl\UserControlCases.Designer.cs">
+      <DependentUpon>UserControlCases.cs</DependentUpon>
     </Compile>
     <Compile Include="UserControl\UserControlClients.cs">
       <SubType>UserControl</SubType>
@@ -78,17 +112,11 @@
     <Compile Include="UserControl\UserControlEmployees.Designer.cs">
       <DependentUpon>UserControlEmployees.cs</DependentUpon>
     </Compile>
-    <Compile Include="UserControl\UserControlCases.cs">
+    <Compile Include="UserControl\UserControlHelp.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="UserControl\UserControlCases.Designer.cs">
-      <DependentUpon>UserControlCases.cs</DependentUpon>
-    </Compile>
-    <Compile Include="UserControl\UserControl1Test.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="UserControl\UserControl1Test.Designer.cs">
-      <DependentUpon>UserControl1Test.cs</DependentUpon>
+    <Compile Include="UserControl\UserControlHelp.Designer.cs">
+      <DependentUpon>UserControlHelp.cs</DependentUpon>
     </Compile>
     <EmbeddedResource Include="GUIForm.resx">
       <DependentUpon>GUIForm.cs</DependentUpon>
@@ -102,8 +130,11 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
-    <EmbeddedResource Include="UserControl\UserControlHelp.resx">
-      <DependentUpon>UserControlHelp.cs</DependentUpon>
+    <EmbeddedResource Include="UserControl\UserControl1Test.resx">
+      <DependentUpon>UserControl1Test.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="UserControl\UserControlCases.resx">
+      <DependentUpon>UserControlCases.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="UserControl\UserControlClients.resx">
       <DependentUpon>UserControlClients.cs</DependentUpon>
@@ -111,11 +142,8 @@
     <EmbeddedResource Include="UserControl\UserControlEmployees.resx">
       <DependentUpon>UserControlEmployees.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="UserControl\UserControlCases.resx">
-      <DependentUpon>UserControlCases.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="UserControl\UserControl1Test.resx">
-      <DependentUpon>UserControl1Test.cs</DependentUpon>
+    <EmbeddedResource Include="UserControl\UserControlHelp.resx">
+      <DependentUpon>UserControlHelp.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="packages.config" />
     <None Include="Properties\Settings.settings">
@@ -132,8 +160,12 @@
     <None Include="App.config" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\BusinessLogic\BusinessLogic.csproj">
-      <Project>{1290da00-3501-4e69-b647-0510e1f68341}</Project>
+    <ProjectReference Include="..\DataAccess\DataAccess.csproj">
+      <Project>{fbcae7dd-44c9-46de-ae32-0e9d30db6032}</Project>
+      <Name>DataAccess</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\LawHouse\BusinessLogic.csproj">
+      <Project>{bd1fa036-c42c-45b8-af18-50a46c49989c}</Project>
       <Name>BusinessLogic</Name>
     </ProjectReference>
   </ItemGroup>
