+++ GUI/packages.config
@@ -1,4 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="ObjectListView.Official" version="2.9.1" targetFramework="net461" />
+  <package id="System.Configuration.ConfigurationManager" version="4.4.1" targetFramework="net461" />
+  <package id="System.Reflection.Emit" version="4.3.0" targetFramework="net461" />
+  <package id="Unity" version="5.8.11" targetFramework="net461" />
 </packages>
