+++ LawHouse.sln
@@ -7,7 +7,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BusinessLogic", "LawHouse\B
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DataAccess", "DataAccess\DataAccess.csproj", "{FBCAE7DD-44C9-46DE-AE32-0E9D30DB6032}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowsFormsApp1", "WindowsFormsApp1\WindowsFormsApp1.csproj", "{2BF80A96-6EAF-4A93-8A5E-9339980CC3BD}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GUI", "GUI\GUI.csproj", "{4A9F7FB1-E6DD-4789-8F0A-153B862FD38D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +23,10 @@ Global
 		{FBCAE7DD-44C9-46DE-AE32-0E9D30DB6032}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FBCAE7DD-44C9-46DE-AE32-0E9D30DB6032}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FBCAE7DD-44C9-46DE-AE32-0E9D30DB6032}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2BF80A96-6EAF-4A93-8A5E-9339980CC3BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2BF80A96-6EAF-4A93-8A5E-9339980CC3BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2BF80A96-6EAF-4A93-8A5E-9339980CC3BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2BF80A96-6EAF-4A93-8A5E-9339980CC3BD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4A9F7FB1-E6DD-4789-8F0A-153B862FD38D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4A9F7FB1-E6DD-4789-8F0A-153B862FD38D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4A9F7FB1-E6DD-4789-8F0A-153B862FD38D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4A9F7FB1-E6DD-4789-8F0A-153B862FD38D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
