--- Model/Entity.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Model
-{
-    public abstract class Entity
-    {
-        public int ID { get; set; }
-    }
-}
