--- WindowsFormsApp1/Program.cs
@@ -1,23 +1,0 @@
-﻿using BusinessLogic;
-using DataAccess;
-using System;
-using System.Windows.Forms;
-using Unity;
-using Unity.Injection;
-
-namespace WindowsFormsApp1
-{
-    static class Program
-    {
-        /// <summary>
-        /// The main entry point for the application.
-        /// </summary>
-        [STAThread]
-        static void Main()
-        {
-            Application.EnableVisualStyles();
-            Application.SetCompatibleTextRenderingDefault(false);
-            Application.Run(new GUIForm());
-        }
-    }
-}
